refact(ContributionFlow): Move helpers to global lib
