+++ lib/__tests__/tier-utils.test.js
@@ -1,4 +1,4 @@
-import { getTierPresets } from '../utils';
+import { getTierPresets } from '../tier-utils';
 
 const getTierWithMin = min => ({ minimumAmount: { valueInCents: min } });
